+++ packages/react-components/src/logo/index.js
@@ -1,0 +1,9 @@
+import LogoHeader from './components/logo-header'
+import LogoFooter from './components/logo-footer'
+
+export default {
+  LogoHeader,
+  LogoFooter,
+}
+
+export { LogoHeader, LogoFooter }